Split in hybrid default+installer recipe
test_package based on work of @madebr

general structure based on the work of @SSE4 and @uilianries
