+++ test_package_installer/basic_nr.txt
@@ -1,0 +1,5 @@
+# this is a comment
+foo = "bar"
+num = 43
+setting = false
+
